requirements.txt
absl-py==2.1.0
alabaster==0.7.16
altair==5.2.0
arrow==1.3.0
astroid==3.2.4
asttokens==2.4.1
astunparse==1.6.3
atomicwrites==1.4.1
attrs==23.2.0
autopep8==2.0.4
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.2.0
bleach==6.1.0
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cookiecutter==2.6.0
cryptography==42.0.4
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
distlib==0.3.8
docstring-to-markdown==0.15
docutils==0.20.1
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
flake8==7.0.0
flatbuffers==23.5.26
fonttools==4.47.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
html5lib==1.1
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.1
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.21.0
isort==5.13.2
jaraco.classes==3.3.1
jedi==0.19.1
jellyfish==1.0.3
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
keras==3.4.1
keras-utils==1.0.13
keyring==24.3.0
kiwisolver==1.4.5
latest==0.6.0
libclang==16.0.6
lxml==5.2.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mlxtend==0.23.1
more-itertools==10.2.0
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
numpy==1.26.3
numpydoc==1.6.0
oauthlib==3.2.2
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.3
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
plotly==5.22.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydeck==0.8.1b0
pydocstyle==6.3.0
pyflakes==3.2.0
pyfunctional==1.4.3
Pygments==2.17.2
pylint==3.2.6
pylint-venv==3.0.3
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.2
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.11.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.3.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rope==1.12.0
rpds-py==0.17.1
rsa==4.9
Rtree==1.2.0
scikeras==0.13.0
scikit-learn==1.5.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sklearn-utils==0.0.15
smmap==5.0.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
spyder==5.5.5
spyder-kernels==2.5.2
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.31.0
style==1.1.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
textdistance==4.6.1
threadpoolctl==3.3.0
three-merge==0.1.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.1
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
update==0.0.1
urllib3==2.2.0
validators==0.22.0
virtualenv==20.26.2
watchdog==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
whatthepatch==1.0.5
wrapt==1.14.1
yapf==0.40.2
zipp==3.17.0
